+++ ykdl/compact.py
@@ -6,9 +6,14 @@ import platform
 import struct
 
 if sys.version_info[0] == 3:
-    from urllib.request import Request, urlopen, HTTPSHandler, build_opener, HTTPCookieProcessor, install_opener, ProxyHandler
-    from urllib.parse import urlencode, urlparse
+    from urllib.request import Request, urlopen, HTTPSHandler, build_opener, HTTPCookieProcessor, install_opener, ProxyHandler, getproxies
+    from urllib.parse import urlencode, urlparse, urlsplit, urljoin, parse_qs
     from http.client import HTTPConnection
+    from http.server import BaseHTTPRequestHandler
+    import socketserver as SocketServer
+    import queue as Queue
+    import _thread as thread
+    from html import unescape
     compact_str = str
     compact_bytes = bytes
     from urllib.parse import unquote as compact_unquote
@@ -24,9 +29,13 @@ if sys.version_info[0] == 3:
         return isinstance(s, str)
 else:
     from urllib2 import Request, urlopen, HTTPSHandler, build_opener, HTTPCookieProcessor, install_opener, ProxyHandler
-    from urllib import urlencode
-    from urlparse import urlparse
+    from urllib import urlencode, getproxies
+    from urlparse import urlparse, urlsplit, urljoin, parse_qs
     from httplib import HTTPConnection
+    from BaseHTTPServer import BaseHTTPRequestHandler
+    import SocketServer
+    import Queue
+    import thread
     import types
     compact_str = unicode
     def compact_bytes(string, encode):
@@ -49,6 +58,22 @@ else:
         return codecs.open(tmp.name, mode, encoding)
     def compact_isstr(s):
         return isinstance(s, types.UnicodeType) or isinstance(s, str)
+    import HTMLParser
+    def unescape(s):
+        html_parser = HTMLParser.HTMLParser()
+        return html_parser.unescape(s)
+
+# Return addrlist sequence at random, it can help create_connection function
+import socket
+import random
+
+def getaddrinfo(*args, **kwargs):
+    addrlist = _getaddrinfo(*args, **kwargs)
+    random.shuffle(addrlist)
+    return addrlist
+
+_getaddrinfo = socket.getaddrinfo
+socket.getaddrinfo = getaddrinfo
 
 try:
     struct.pack('!I', 0)