+++ ykdl/extractor.py
@@ -15,12 +15,11 @@ class VideoExtractor():
         self.__init__()
         if compact_isstr(url) and url.startswith('http'):
             self.url = url
+            if self.list_only():
+                return self.parser_list(url)
         else:
             self.vid= url
 
-        # if info is returned by prepare, no need go extractor
-        # else go extractor.
-        # info is instance of VideoInfo
         info = self.prepare()
         return info
 
@@ -32,11 +31,19 @@ class VideoExtractor():
         for video in video_list:
             yield self.parser(video)
 
+    def __getattr__(self, attr):
+        return None
+
     def prepare(self):
         pass
 
-    def extractor(self):
+    def prepare_list(self):
         pass
 
-    def prepare_list(self):
+    def list_only(self):
+        """
+        this API is to check if only the list informations is included
+        if true, go to parser list mode
+        MUST override!!
+        """
         pass