+++ ykdl/extractors/tudou.py
@@ -2,18 +2,14 @@
 # -*- coding: utf-8 -*-
 
 from .youku import Youku
-from ykdl.util.html import get_location
-import re
+from ykdl.util.html import get_location, add_header
+from ykdl.util.match import match1
 
 class Tudou(Youku):
     name = u"Tudou (土豆)"
 
-    def __init__(self):
-        Youku.__init__(self)
-        self.ccode = '0402'
-
     def prepare(self):
-        if not re.search('video.tuodou.com', self.url):
+        if match1(self.url, '(new-play|video)\.tudou\.com/') is None:
             self.url = get_location(self.url)
         return Youku.prepare(self)